Rename metadata.ts to Metadata.ts
